refactor: rename runner to runner-worker
